requirements.txt
fastapi
uvicorn
qdrant-client
neo4j
sentence-transformers
transformers
torch
torchvision
pillow
easyocr
unstructured[pdf]
llama-index-core
mistralai
pydantic